Saved some more bits switching to ints

+++ ssd1306_minimal/NexStarGPS.h
@@ -99,7 +99,7 @@ class NexstarMessageSender
 public:
 	NexstarMessageSender(TinyGPS* _gps);
 	nexstar_msg_union* getMessage();
-	bool send(soss* serial, unsigned long& sentByteCount);
+	bool send(soss* serial, unsigned int& sentByteCount);
 	bool handleMessage(NexstarMessageReceiver* receiver);
 
 protected:
